[REFACTOR] : useStatus 추가로 코드 중복 제거
