+++ src/components/common/Modal/index.tsx
@@ -11,12 +11,12 @@ interface ModalProps {
   children: ReactNode;
 }
 
-interface VariantType {
+interface VariantTypes {
   default: string;
   primary: string;
 }
 
-const VARIANTS: VariantType = {
+const VARIANTS: VariantTypes = {
   default: styles.modal_section__default,
   primary: styles.modal_section__primary,
 };
@@ -59,7 +59,7 @@ export default function Modal({
     >
       <section
         className={classNames(
-          VARIANTS[variant as keyof VariantType],
+          VARIANTS[variant as keyof VariantTypes],
           isModalOpen
             ? `${styles.modal_section} ${styles.modal_section__opened}`
             : `${styles.modal_section} ${styles.modal_section__closed}`,
