+++ src/pages/chat/MainChat/index.tsx
@@ -7,7 +7,6 @@ import Button from '@/components/common/Button';
 import ChatLoading from '@/components/chat/ChatLoading';
 import MarkDown from '@/components/chat/MarkDown';
 import { API_URLS, BASE_URL, ROUTE_PATHS } from '@/constants/config';
-import Header from '@/components/common/Header';
 import ImageUploadButton from '@/components/chat/ImageUploadButton';
 import gildong from '@/assets/gildong_3d.png';
 import AddItineraryButton from '@/components/travel/AddItineraryButton';
@@ -24,6 +23,7 @@ import mountain from '@/assets/mountain.png';
 import wheelchair from '@/assets/wheelchair.png';
 import voiceLoading from '@/assets/loading_voice.gif';
 import useRecording from '@/hooks/useRecording';
+import useStatus from '@/hooks/useStatus';
 import styles from './styles.module.scss';
 interface ChatProps {
   home?: boolean;
@@ -52,6 +52,7 @@ export default function MainChat({ home }: ChatProps) {
     startRecording,
     stopRecording,
   } = useRecording(setValue);
+  useStatus('mainChat', 'AI Travel Planner 길동이');
 
   const handleInput = (event: React.ChangeEvent<HTMLInputElement>) => {
     setValue(event.target.value);
@@ -172,51 +173,148 @@ export default function MainChat({ home }: ChatProps) {
   }, [answer, list]);
 
   return (
-    <>
-      <Header>AI Travel Planner 길동이</Header>
-      <div className={styles.pageWrapper}>
-        <div className={styles.content}>
-          <div
-            className={styles.questionContainer}
-            ref={(el) => (scrollRef.current[1] = el)}
-            onWheel={handleWheel}
-          >
-            {!home ? (
-              <div className={styles.questionBackground}>
-                <div ref={(el) => (scrollRef.current[2] = el)}>
-                  {list?.map((el, index) => (
-                    <div
-                      className={styles.question}
-                      key={index}
-                      ref={(el) => {
-                        scrollRef.current[0] = el;
-                      }}
-                    >
-                      {el.image_name ? (
-                        <div className={styles.uploadImage}>
-                          <img
-                            src={`${BASE_URL}${API_URLS.viewImage}${el.image_name}`}
-                          />
+    <div className={styles.pageWrapper}>
+      <div className={styles.content}>
+        <div
+          className={styles.questionContainer}
+          ref={(el) => (scrollRef.current[1] = el)}
+          onWheel={handleWheel}
+        >
+          {!home ? (
+            <div className={styles.questionBackground}>
+              <div ref={(el) => (scrollRef.current[2] = el)}>
+                {list?.map((el, index) => (
+                  <div
+                    className={styles.question}
+                    key={index}
+                    ref={(el) => {
+                      scrollRef.current[0] = el;
+                    }}
+                  >
+                    {el.image_name ? (
+                      <div className={styles.uploadImage}>
+                        <img
+                          src={`${BASE_URL}${API_URLS.viewImage}${el.image_name}`}
+                        />
+                      </div>
+                    ) : null}
+                    {el.question}
+                    <div className={styles.answer}>
+                      <MarkDown text={el.answer} />
+                      {el.itinerary_id ? (
+                        <div className={styles.addButton}>
+                          <AddItineraryButton id={el.itinerary_id} />
                         </div>
                       ) : null}
-                      {el.question}
-                      <div className={styles.answer}>
-                        <MarkDown text={el.answer} />
-                        {el.itinerary_id ? (
-                          <div className={styles.addButton}>
-                            <AddItineraryButton id={el.itinerary_id} />
-                          </div>
-                        ) : null}
-                      </div>
                     </div>
-                  ))}
-                  {question || (image && !isImageOpen) ? (
-                    <div
-                      className={styles.question}
-                      ref={(el) => {
-                        scrollRef.current[0] = el;
-                      }}
-                    >
+                  </div>
+                ))}
+                {question || (image && !isImageOpen) ? (
+                  <div
+                    className={styles.question}
+                    ref={(el) => {
+                      scrollRef.current[0] = el;
+                    }}
+                  >
+                    {image ? (
+                      <div className={styles.uploadImage}>
+                        <img src={`${BASE_URL}${API_URLS.viewImage}${image}`} />
+                      </div>
+                    ) : null}
+                    {question}
+                    <div className={styles.answer}>
+                      {isChatLoading ? (
+                        <ChatLoading />
+                      ) : (
+                        <MarkDown text={answer} />
+                      )}
+                    </div>
+                  </div>
+                ) : null}
+              </div>
+              {stop ? null : <div className={styles.margin}></div>}
+            </div>
+          ) : (
+            <div className={styles.default}>
+              <div className={styles.title}>
+                길동이에게 여행 일정을 맡겨보세요!
+              </div>
+              <img src={gildong} className={styles.img} />
+              <div className={styles.exampleWrapper}>
+                <div
+                  className={styles.example}
+                  onClick={() => {
+                    fetchSSE('서울 근처에 등산하기 좋은 장소 추천해줄래?');
+                    navigate(ROUTE_PATHS.mainChat);
+                  }}
+                >
+                  <span className={styles.text}>
+                    서울 근처에 등산하기 좋은 장소 추천해줄래?
+                  </span>
+                  <div className={styles.homeImage}>
+                    <img src={mountain} />
+                  </div>
+                </div>
+                <div
+                  className={styles.example}
+                  onClick={() => {
+                    fetchSSE(
+                      '부산에서 휠체어 이용이 편리한 2박 3일 여행코스를 알려줄래?',
+                    );
+                    navigate(ROUTE_PATHS.mainChat);
+                  }}
+                >
+                  <span className={styles.text}>
+                    부산에서 휠체어 이용이 편리한 2박 3일 여행코스를 알려줄래?
+                  </span>
+                  <div className={styles.homeImage}>
+                    <img src={wheelchair} />
+                  </div>
+                </div>
+                <div
+                  className={styles.example}
+                  onClick={() => {
+                    fetchSSE(
+                      ' 통영에서 시각장애인도 편하게 여행할 수 있는 1박 2일 코스 추천해줄 수 있을까?',
+                    );
+                    navigate(ROUTE_PATHS.mainChat);
+                  }}
+                >
+                  <span className={styles.text}>
+                    통영에서 시각장애인도 편하게 여행할 수 있는 1박 2일 코스
+                    추천해줄 수 있을까?
+                  </span>
+                  <div className={styles.homeImage}>
+                    <img src={blindPerson} />
+                  </div>
+                </div>
+                <div
+                  className={styles.example}
+                  onClick={() => {
+                    setImage('20231020003408_바다사진.jpg');
+                    fetchSSE(
+                      '내가 보내준 이미지와 비슷한 분위기의 여행지를 알려줄래?',
+                      '20231020003408_바다사진.jpg',
+                    );
+                    navigate(ROUTE_PATHS.mainChat);
+                  }}
+                >
+                  <span className={styles.text}>
+                    내가 보내준 이미지와 비슷한 분위기의 여행지를 알려줄래?
+                  </span>
+                  <div className={styles.homeImage}>
+                    <img src={beach} />
+                  </div>
+                </div>
+              </div>
+            </div>
+          )}
+          <div className={styles.chatContainer}>
+            <div className={styles.chatWrapper}>
+              <div className={styles.chat}>
+                {isImageOpen ? (
+                  <div className={styles.imageContainer}>
+                    <div className={styles.uploadImage}>
                       {image ? (
                         <div className={styles.uploadImage}>
                           <img
@@ -224,173 +322,71 @@ export default function MainChat({ home }: ChatProps) {
                           />
                         </div>
                       ) : null}
-                      {question}
-                      <div className={styles.answer}>
-                        {isChatLoading ? (
-                          <ChatLoading />
-                        ) : (
-                          <MarkDown text={answer} />
-                        )}
+                      <span className={styles.cancel}>
+                        <img
+                          src={remove}
+                          onClick={() => {
+                            setImage('');
+                            setIsImageOpen(false);
+                          }}
+                        />
+                      </span>
+                    </div>
+                  </div>
+                ) : null}
+                {isRecording ? (
+                  <div className={styles.voiceContainer}>
+                    <div className={styles.voice}>
+                      듣고 있습니다.
+                      <div className={styles.pulse} onClick={stopRecording}>
+                        <div className={styles.rectangle} />
                       </div>
                     </div>
-                  ) : null}
+                  </div>
+                ) : isSTTLoading ? (
+                  <div className={styles.voiceContainer}>
+                    <div className={styles.voice}>
+                      <img src={voiceLoading} />
+                    </div>
+                  </div>
+                ) : null}
+                <div className={styles.icon}>
+                  <ImageUploadButton setIsImageOpen={setIsImageOpen} />
                 </div>
-                {stop ? null : <div className={styles.margin}></div>}
-              </div>
-            ) : (
-              <div className={styles.default}>
-                <div className={styles.title}>
-                  길동이에게 여행 일정을 맡겨보세요!
-                </div>
-                <img src={gildong} className={styles.img} />
-                <div className={styles.exampleWrapper}>
-                  <div
-                    className={styles.example}
-                    onClick={() => {
-                      fetchSSE('서울 근처에 등산하기 좋은 장소 추천해줄래?');
-                      navigate(ROUTE_PATHS.mainChat);
-                    }}
-                  >
-                    <span className={styles.text}>
-                      서울 근처에 등산하기 좋은 장소 추천해줄래?
-                    </span>
-                    <div className={styles.homeImage}>
-                      <img src={mountain} />
-                    </div>
-                  </div>
-                  <div
-                    className={styles.example}
-                    onClick={() => {
-                      fetchSSE(
-                        '부산에서 휠체어 이용이 편리한 2박 3일 여행코스를 알려줄래?',
-                      );
-                      navigate(ROUTE_PATHS.mainChat);
-                    }}
-                  >
-                    <span className={styles.text}>
-                      부산에서 휠체어 이용이 편리한 2박 3일 여행코스를 알려줄래?
-                    </span>
-                    <div className={styles.homeImage}>
-                      <img src={wheelchair} />
-                    </div>
-                  </div>
-                  <div
-                    className={styles.example}
-                    onClick={() => {
-                      fetchSSE(
-                        ' 통영에서 시각장애인도 편하게 여행할 수 있는 1박 2일 코스 추천해줄 수 있을까?',
-                      );
-                      navigate(ROUTE_PATHS.mainChat);
-                    }}
-                  >
-                    <span className={styles.text}>
-                      통영에서 시각장애인도 편하게 여행할 수 있는 1박 2일 코스
-                      추천해줄 수 있을까?
-                    </span>
-                    <div className={styles.homeImage}>
-                      <img src={blindPerson} />
-                    </div>
-                  </div>
-                  <div
-                    className={styles.example}
-                    onClick={() => {
-                      setImage('20231020003408_바다사진.jpg');
-                      fetchSSE(
-                        '내가 보내준 이미지와 비슷한 분위기의 여행지를 알려줄래?',
-                        '20231020003408_바다사진.jpg',
-                      );
-                      navigate(ROUTE_PATHS.mainChat);
-                    }}
-                  >
-                    <span className={styles.text}>
-                      내가 보내준 이미지와 비슷한 분위기의 여행지를 알려줄래?
-                    </span>
-                    <div className={styles.homeImage}>
-                      <img src={beach} />
-                    </div>
-                  </div>
-                </div>
-              </div>
-            )}
-            <div className={styles.chatContainer}>
-              <div className={styles.chatWrapper}>
-                <div className={styles.chat}>
-                  {isImageOpen ? (
-                    <div className={styles.imageContainer}>
-                      <div className={styles.uploadImage}>
-                        {image ? (
-                          <div className={styles.uploadImage}>
-                            <img
-                              src={`${BASE_URL}${API_URLS.viewImage}${image}`}
-                            />
-                          </div>
-                        ) : null}
-                        <span className={styles.cancel}>
-                          <img
-                            src={remove}
-                            onClick={() => {
-                              setImage('');
-                              setIsImageOpen(false);
-                            }}
-                          />
-                        </span>
-                      </div>
-                    </div>
+                <input
+                  className={styles.input}
+                  onChange={handleInput}
+                  value={value}
+                  onKeyDown={handleEnter}
+                  placeholder="무엇이든 물어보세요!"
+                />
+                <div className={styles.send}>
+                  {value || isImageOpen ? (
+                    <Button
+                      size="sm"
+                      color="white"
+                      icon={<TbSend />}
+                      iconBtn={true}
+                      variant="primary"
+                      onClick={handleSubmit}
+                    />
+                  ) : !isRecordingStarting ? (
+                    <Button
+                      icon={<BsFillMicFill />}
+                      size="sm"
+                      color="secondary"
+                      iconBtn={true}
+                      onClick={startRecording}
+                    >
+                      녹음 시작
+                    </Button>
                   ) : null}
-                  {isRecording ? (
-                    <div className={styles.voiceContainer}>
-                      <div className={styles.voice}>
-                        듣고 있습니다.
-                        <div className={styles.pulse} onClick={stopRecording}>
-                          <div className={styles.rectangle} />
-                        </div>
-                      </div>
-                    </div>
-                  ) : isSTTLoading ? (
-                    <div className={styles.voiceContainer}>
-                      <div className={styles.voice}>
-                        <img src={voiceLoading} />
-                      </div>
-                    </div>
-                  ) : null}
-                  <div className={styles.icon}>
-                    <ImageUploadButton setIsImageOpen={setIsImageOpen} />
-                  </div>
-                  <input
-                    className={styles.input}
-                    onChange={handleInput}
-                    value={value}
-                    onKeyDown={handleEnter}
-                    placeholder="무엇이든 물어보세요!"
-                  />
-                  <div className={styles.send}>
-                    {value || isImageOpen ? (
-                      <Button
-                        size="sm"
-                        color="white"
-                        icon={<TbSend />}
-                        iconBtn={true}
-                        variant="primary"
-                        onClick={handleSubmit}
-                      />
-                    ) : !isRecordingStarting ? (
-                      <Button
-                        icon={<BsFillMicFill />}
-                        size="sm"
-                        color="secondary"
-                        iconBtn={true}
-                        onClick={startRecording}
-                      >
-                        녹음 시작
-                      </Button>
-                    ) : null}
-                  </div>
                 </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-    </>
+    </div>
   );
 }
