+++ src/pages/travel/TravelDetails/index.tsx
@@ -10,15 +10,15 @@ import { RiGuideFill } from 'react-icons/ri';
 import { MdOutlineOndemandVideo } from 'react-icons/md';
 import { BsFillSignpostSplitFill } from 'react-icons/bs';
 import { getTravelDetailAPI } from '@/services/travel';
-import { detailType } from '@/types/travel';
-import Header from '@/components/common/Header';
+import { detailTypes } from '@/types/travel';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
+import useStatus from '@/hooks/useStatus';
 import styles from './styles.module.scss';
 
 export default function TravelDetails() {
   const { id } = useParams();
   const [currentTab, setCurrentTab] = useState(0);
-  const [detailData, setDetailData] = useState<detailType>({
+  const [detailData, setDetailData] = useState<detailTypes>({
     title: '',
     overview: '',
     contenttypeid: '',
@@ -36,6 +36,7 @@ export default function TravelDetails() {
     { name: '지체장애인', content: detailData.physical },
     { name: '청각장애인', content: detailData.hearing },
   ];
+  useStatus('travelDetails', '');
 
   const getTravelDetail = async (id: string) => {
     setIsLoading(true);
@@ -53,148 +54,145 @@ export default function TravelDetails() {
   }, [id]);
 
   return (
-    <>
-      <Header back={true} />
-      <div className={styles.pageWrapper}>
-        {isLoading ? (
-          <div className={styles.loading}>
-            <LoadingSpinner />
-          </div>
-        ) : (
-          <>
-            <div className={styles.title}>{detailData.title}</div>
-            {detailData?.url ? (
-              <img src={detailData.url} className={styles.img} />
+    <div className={styles.pageWrapper}>
+      {isLoading ? (
+        <div className={styles.loading}>
+          <LoadingSpinner />
+        </div>
+      ) : (
+        <>
+          <div className={styles.title}>{detailData.title}</div>
+          {detailData?.url ? (
+            <img src={detailData.url} className={styles.img} />
+          ) : null}
+          <div className={styles.contentWrapper}>
+            {detailData?.addr ? (
+              <div className={styles.content}>
+                <FaMapMarkerAlt /> {detailData.addr}
+              </div>
             ) : null}
-            <div className={styles.contentWrapper}>
-              {detailData?.addr ? (
-                <div className={styles.content}>
-                  <FaMapMarkerAlt /> {detailData.addr}
-                </div>
-              ) : null}
-              {detailData?.tel ? (
-                <div className={styles.content}>
-                  <BsFillTelephoneFill /> {detailData.tel}
-                </div>
-              ) : null}
-              {detailData?.overview ? (
-                <p className={styles.info}>
-                  {detailData.overview.replaceAll('<br>', '')}
-                </p>
-              ) : null}
-              {detailData?.visual &&
-              detailData?.physical &&
-              detailData?.hearing ? (
-                <>
-                  <ul className={styles.tabWrapper}>
-                    {tabArr.map((el, index) => (
-                      <li
-                        key={index}
-                        className={
-                          index === currentTab
-                            ? `${styles.tab} ${styles.focused}`
-                            : styles.tab
-                        }
-                        onClick={() => setCurrentTab(index)}
-                      >
-                        {el.name}
-                      </li>
-                    ))}
-                  </ul>
-                  <div className={styles.tabContent}>
-                    {tabArr[currentTab].content?.split('#').map((el, index) => {
-                      if (el.includes('주차')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <AiFillCar />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('점자')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <GrBraille />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('휠체어')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <FaWheelchair />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('엘리베이터')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <MdElevator />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('화장실')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <FaRestroom />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('보조견')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <FaDog />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('음성')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <AiFillSound />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('안내요원')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <BsFillPersonFill />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('안내 시스템')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <RiGuideFill />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('표지')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <BsFillSignpostSplitFill />
-                            {el}
-                          </div>
-                        );
-                      } else if (el.includes('비디오')) {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <MdOutlineOndemandVideo />
-                            {el}
-                          </div>
-                        );
-                      } else {
-                        return (
-                          <div className={styles.iconWrapper} key={index}>
-                            <AiOutlineCheck /> {el}
-                          </div>
-                        );
+            {detailData?.tel ? (
+              <div className={styles.content}>
+                <BsFillTelephoneFill /> {detailData.tel}
+              </div>
+            ) : null}
+            {detailData?.overview ? (
+              <p className={styles.info}>
+                {detailData.overview.replaceAll('<br>', '')}
+              </p>
+            ) : null}
+            {detailData?.visual &&
+            detailData?.physical &&
+            detailData?.hearing ? (
+              <>
+                <ul className={styles.tabWrapper}>
+                  {tabArr.map((el, index) => (
+                    <li
+                      key={index}
+                      className={
+                        index === currentTab
+                          ? `${styles.tab} ${styles.focused}`
+                          : styles.tab
                       }
-                    })}
-                  </div>
-                </>
-              ) : null}
-            </div>
-          </>
-        )}
-      </div>
-    </>
+                      onClick={() => setCurrentTab(index)}
+                    >
+                      {el.name}
+                    </li>
+                  ))}
+                </ul>
+                <div className={styles.tabContent}>
+                  {tabArr[currentTab].content?.split('#').map((el, index) => {
+                    if (el.includes('주차')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <AiFillCar />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('점자')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <GrBraille />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('휠체어')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <FaWheelchair />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('엘리베이터')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <MdElevator />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('화장실')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <FaRestroom />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('보조견')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <FaDog />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('음성')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <AiFillSound />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('안내요원')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <BsFillPersonFill />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('안내 시스템')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <RiGuideFill />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('표지')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <BsFillSignpostSplitFill />
+                          {el}
+                        </div>
+                      );
+                    } else if (el.includes('비디오')) {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <MdOutlineOndemandVideo />
+                          {el}
+                        </div>
+                      );
+                    } else {
+                      return (
+                        <div className={styles.iconWrapper} key={index}>
+                          <AiOutlineCheck /> {el}
+                        </div>
+                      );
+                    }
+                  })}
+                </div>
+              </>
+            ) : null}
+          </div>
+        </>
+      )}
+    </div>
   );
 }
