+++ src/pages/user/SignIn/index.tsx
@@ -2,10 +2,12 @@ import { useEffect } from 'react';
 import gildong from '@/assets/gildong_3d_bg.png';
 import kakao from '@/assets/kakao.png';
 import { JAVASCRIPT_KEY, REDIRECT_URL } from '@/constants/auth';
-import Header from '@/components/common/Header';
+import useStatus from '@/hooks/useStatus';
 import styles from './styles.module.scss';
 
 export default function SignIn() {
+  useStatus('signIn', '');
+
   useEffect(() => {
     const script = document.createElement('script');
     script.src = 'https://t1.kakaocdn.net/kakao_js_sdk/2.4.0/kakao.min.js';
@@ -27,23 +29,20 @@ export default function SignIn() {
   };
 
   return (
-    <>
-      <Header />
-      <div className={styles.pageWrapper}>
-        <div className={styles.contentWrapper}>
-          <title className={styles.titleWrapper}>
-            <div className={styles.text}>대화로 만들어가는 여행플래너</div>
-            <div className={styles.title}>AI Travel Planner 길동이</div>
-          </title>
-          <img src={gildong} className={styles.img} />
-          <button className={styles.btnWrapper} onClick={loginWithKakao}>
-            <div className={styles.btn}>
-              <img src={kakao} />
-              <span>카카오 로그인</span>
-            </div>
-          </button>
-        </div>
+    <div className={styles.pageWrapper}>
+      <div className={styles.contentWrapper}>
+        <title className={styles.titleWrapper}>
+          <div className={styles.text}>대화로 만들어가는 여행플래너</div>
+          <div className={styles.title}>AI Travel Planner 길동이</div>
+        </title>
+        <img src={gildong} className={styles.img} />
+        <button className={styles.btnWrapper} onClick={loginWithKakao}>
+          <div className={styles.btn}>
+            <img src={kakao} />
+            <span>카카오 로그인</span>
+          </div>
+        </button>
       </div>
-    </>
+    </div>
   );
 }
