+++ src/store/atom/globalAtom.ts
@@ -1,0 +1,11 @@
+import { atom } from 'recoil';
+
+const headerStatusState = atom<HeaderStatusTypes>({
+  key: 'headerStatus',
+  default: {
+    pageName: 'AI Travel Planner 길동이',
+    title: '',
+  },
+});
+
+export { headerStatusState };
