+++ src/types/global.d.ts
@@ -1,7 +1,8 @@
-declare interface ErrorResponse {
-  message: string;
-}
-
 declare interface Window {
   Kakao: any;
+}
+
+declare interface HeaderStatusTypes {
+  pageName: string;
+  title: string | undefined;
 }
